readme.txt

+++ lr1/src/main/java/org/example/Main.java
@@ -1,6 +1,5 @@
 package org.example;
 
-import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.PrintWriter;
 import java.util.Scanner;
@@ -18,7 +17,7 @@ public class Main {
         double numerator = Math.pow(Math.cos(y),3)+Math.pow(2,x)*d;
         double denominator = Math.pow(Math.E,y)+Math.log(Math.pow(Math.sin(x),2)+7.4);
         double result = numerator/denominator;
-        System.out.println("Result is: "+(int)result);
+        System.out.println("Result is: "+result);
         int alphabetNumber = (int) (result*9); // 9 - номер варианта, на который надо умножить
         System.out.println("Number in alphabet: "+alphabetNumber);
         String str = String.valueOf(alphabetNumber); // Приведение числа к строке
